+++ cli/lib/generators/class/copy_with.dart
@@ -3,18 +3,20 @@ import '../swagger/table.dart';
 String generateCopyWithMethod(String className, Table table) {
   final columns = table.columns;
   final StringBuffer code = StringBuffer();
-
+  //  static const _unset = Object();
+  code.writeln('static const _unset = Object();');
   code.writeln('${className} copyWith({');
 
   columns.forEach((columnName, columnDetails) {
-    code.writeln('${columnDetails.dartType}? $columnName,');
+    code.writeln('Object? $columnName = _unset,');
   });
 
   code.writeln('}) {');
 
   code.writeln('return $className(');
   columns.forEach((columnName, columnDetails) {
-    code.writeln('$columnName: $columnName ?? this.$columnName,');
+    code.writeln(
+        '$columnName: $columnName == _unset ? this.$columnName : $columnName as ${columnDetails.dartType}${columnDetails.isNullable ? "?" : ""},');
   });
 
   code.writeln(');');